Update README.md
This project is aimed at developing a web based Leave Management Tool, which is of importance to either an organization or a college. This is an Intranet based application that can be accessed throughout the organization or a specified group/Dept. This system can be used to automate the workflow of leave applications and their approvals. There are features like cancellation of leave, approval of leave, report generators etc. in this Tool.